moved wind turbine out of hydraulics
