+++ fbw-common/src/wasm/systems/systems/src/lib.rs
@@ -20,3 +20,4 @@ pub mod pneumatic;
 pub mod pressurization;
 pub mod shared;
 pub mod simulation;
+pub mod wind_turbine;
